Extract Key Pair Handling as separate component
